Add files via upload
A simple program allowing you to find a suitable nonce creating a sha256 hash with the desired amount of 0 for a given input. 

There is a multithreaded version of this. This doesn't work in python though. Making it faster using NumPy with JiT using Numba was not successful. 

As a workaround, I cut it down to hashingSimple.py, made it start at random nonces, compiled it, and ran 8 versions of it to work on an Intel 3610QM Quadcore laptop processor as seen in MultiCoreWorkarround.png . It gave me results after two days.